spotbugs

+++ src/main/java/io/cryostat/recordings/RecordingHelper.java
@@ -1038,9 +1038,7 @@ public class RecordingHelper {
                 .run(
                         () ->
                                 ArchivedRecordingInfo.of(
-                                                recording.target.jvmId,
-                                                filename,
-                                                recording.id.longValue())
+                                                recording.target.jvmId, filename, recording.id)
                                         .persist());
         ArchivedRecording archivedRecording =
                 getArchivedRecordingInfo(recording.target.jvmId, filename).orElseThrow();
